Update strings.xml
Hi,
several times I have corrected the wrong automatic translation that was added to the app, but it still has not been used. Why is that?

This whole translation is mine and we advertise it also on our website dedicated to app translations http://www.droid.sk/

When you release the app with a bad automatic translation it makes us look incompetent. Please use my translation instead of the automatic one.

Many thanks,
Lorant